+++ examples/interconnect_type.go
@@ -3,7 +3,6 @@ package main
 import (
 	"fmt"
 	"github.com/HewlettPackard/oneview-golang/ov"
-	"github.com/HewlettPackard/oneview-golang/utils"
 	"os"
 )
 
@@ -18,7 +17,7 @@ func main() {
 		os.Getenv("ONEVIEW_OV_DOMAIN"),
 		os.Getenv("ONEVIEW_OV_ENDPOINT"),
 		false,
-		2000,
+		2200,
 		"*")
 
 	fmt.Println("#................... Interconnect Type by Name ...............#")
@@ -29,9 +28,9 @@ func main() {
 		fmt.Println(interconnect)
 	}
 
-	interconnect_type := utils.NewNstring("/rest/interconnect-types/a3af4d5e-7114-4e7a-a6c4-f97b707ec87c")
+
 	fmt.Println("#................... Interconnect Type by Uri ....................#")
-	int_uri, err := ovc.GetInterconnectTypeByUri(interconnect_type)
+	int_uri, err := ovc.GetInterconnectTypeByUri(interconnect.URI)
 	if err != nil {
 		fmt.Println(err)
 	} else {