+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
+# [v1.7.0](Unreleased)
+### Notes
+- This release extends supports of the SDK to Oneview API2200.
+
+### Oneview Features supported
+- Interconnects
+- Interconnect Types
+- Logical Interconnect
+- Server Hardware
+- Server Hardware Types
+
 # [v1.6.0] 
 ### Notes
 - This release extends supports of the SDK to Oneview API2000.